+++ plataforma_web/app.py
@@ -11,33 +11,29 @@ from fastapi_pagination import add_pagination
 from config.settings import ACCESS_TOKEN_EXPIRE_MINUTES
 from lib.database import get_db
 
-from plataforma_web.autoridades.views import router as autoridades
-from plataforma_web.distritos.views import router as distritos
-from plataforma_web.listas_de_acuerdos.views import router as listas_de_acuerdos
-from plataforma_web.listas_de_acuerdos_acuerdos.views import router as listas_de_acuerdos_acuerdos
-from plataforma_web.materias.views import router as materias
-from plataforma_web.roles.views import router as roles
-from plataforma_web.usuarios.views import router as usuarios
+from plataforma_web.v1.autoridades.paths import router as v1_autoridades
+from plataforma_web.v1.distritos.paths import router as v1_distritos
+from plataforma_web.v1.materias.paths import router as v1_materias
+from plataforma_web.v1.roles.paths import router as v1_roles
+from plataforma_web.v1.usuarios.paths import router as v1_usuarios
 
-from plataforma_web.usuarios.authentications import (
+from plataforma_web.v1.usuarios.authentications import (
     authenticate_user,
     create_access_token,
     get_current_active_user,
     oauth2_scheme,
 )
-from plataforma_web.usuarios.schemas import Token, UsuarioEnBD
-
-VERSION = "v1.0"
+from plataforma_web.v1.usuarios.schemas import Token, UsuarioInBD
 
 app = FastAPI()
 
-app.include_router(autoridades, prefix=f"/api/{VERSION}/autoridades")
-app.include_router(distritos, prefix=f"/api/{VERSION}/distritos")
-app.include_router(listas_de_acuerdos, prefix=f"/api/{VERSION}/listas_de_acuerdos")
-app.include_router(listas_de_acuerdos_acuerdos, prefix=f"/api/{VERSION}/listas_de_acuerdos_acuerdos")
-app.include_router(materias, prefix=f"/api/{VERSION}/materias")
-app.include_router(roles, prefix=f"/api/{VERSION}/roles")
-app.include_router(usuarios, prefix=f"/api/{VERSION}/usuarios")
+app.include_router(v1_autoridades, prefix=f"/v1/autoridades")
+app.include_router(v1_distritos, prefix=f"/v1/distritos")
+# app.include_router(listas_de_acuerdos, prefix=f"/v1/listas_de_acuerdos")
+# app.include_router(listas_de_acuerdos_acuerdos, prefix=f"/v1/listas_de_acuerdos_acuerdos")
+app.include_router(v1_materias, prefix=f"/v1/materias")
+app.include_router(v1_roles, prefix=f"/v1/roles")
+app.include_router(v1_usuarios, prefix=f"/v1/usuarios")
 
 add_pagination(app)
 
@@ -64,7 +60,7 @@ async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(
     return {"access_token": access_token, "token_type": "bearer"}
 
 
-@app.get("/usuarios/yo/", response_model=UsuarioEnBD)
-async def read_users_me(current_user: UsuarioEnBD = Depends(get_current_active_user)):
+@app.get("/usuarios/yo/", response_model=UsuarioInBD)
+async def read_users_me(current_user: UsuarioInBD = Depends(get_current_active_user)):
     """Mostrar el perfil del usuario"""
     return current_user
