--- plataforma_web/listas_de_acuerdos_acuerdos/schemas.py
@@ -1,30 +1,0 @@
-"""
-Listas de Acuerdos Acuerdos, esquemas de pydantic
-"""
-from datetime import date
-from typing import Optional
-from pydantic import BaseModel
-
-
-class ListaDeAcuerdoAcuerdoNew(BaseModel):
-    """Acuerdo de Listas de Acuerdos nuevo"""
-
-    lista_de_acuerdo_id: int
-    folio: Optional[str] = ""
-    expediente: Optional[str] = ""
-    actor: str
-    demandado: str
-    tipo_acuerdo: str
-    tipo_juicio: str
-    referencia: int
-
-
-class ListaDeAcuerdoAcuerdo(ListaDeAcuerdoAcuerdoNew):
-    """Acuerdo de Listas de Acuerdos"""
-
-    id: int
-    distrito_id: int
-    distrito: str
-    autoridad_id: int
-    autoridad: str
-    fecha: date
