--- plataforma_web/materias/schemas.py
@@ -1,11 +1,0 @@
-"""
-Materias, esquemas de pydantic
-"""
-from pydantic import BaseModel
-
-
-class Materia(BaseModel):
-    """Esquema para materias"""
-
-    id: int
-    materia: str
