--- plataforma_web/roles/schemas.py
@@ -1,11 +1,0 @@
-"""
-Roles, esquemas de pydantic
-"""
-from pydantic import BaseModel
-
-
-class Rol(BaseModel):
-    """Esquema para roles"""
-
-    id: int
-    rol: str
