--- plataforma_web/v1.0/autoridades/crud.py
@@ -1,4 +1,0 @@
-"""
-Autoridades v1.0, CRUD (create, read, update, and delete)
-"""
-from sqlalchemy.orm import Session
