--- plataforma_web/v1.0/autoridades/paths.py
@@ -1,3 +1,0 @@
-"""
-Autoridades v1.0, rutas (paths)
-"""
