--- plataforma_web/v1.0/autoridades/schemas.py
@@ -1,4 +1,0 @@
-"""
-Autoridades v1.0, esquemas
-"""
-from pydantic import BaseModel
