--- plataforma_web/v1.0/distritos/paths.py
@@ -1,3 +1,0 @@
-"""
-Distritos v1.0, rutas (paths)
-"""
