--- plataforma_web/v1.0/distritos/schemas.py
@@ -1,4 +1,0 @@
-"""
-Distritos v1.0, esquemas
-"""
-from pydantic import BaseModel
