--- plataforma_web/v1.0/materias/crud.py
@@ -1,4 +1,0 @@
-"""
-Materias v1.0, CRUD (create, read, update, and delete)
-"""
-from sqlalchemy.orm import Session
