--- plataforma_web/v1.0/materias/paths.py
@@ -1,3 +1,0 @@
-"""
-Materias v1.0, rutas (paths)
-"""
