--- plataforma_web/v1.0/materias/schemas.py
@@ -1,4 +1,0 @@
-"""
-Materias v1.0, esquemas
-"""
-from pydantic import BaseModel
