+++ plataforma_web/v1/distritos/schemas.py
@@ -1,0 +1,17 @@
+"""
+Distritos v1.0, esquemas
+"""
+from pydantic import BaseModel
+
+
+class DistritoOut(BaseModel):
+    """Esquema para entregar distrito"""
+
+    id: int
+    nombre: str
+    nombre_corto: str
+    es_distrito_judicial: bool
+
+    class Config:
+        """SQLAlchemy config"""
+        orm_mode = True
