+++ plataforma_web/v1/materias/paths.py
@@ -1,0 +1,44 @@
+"""
+Materias v1.0, rutas (paths)
+"""
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi_pagination import LimitOffsetPage
+from fastapi_pagination.ext.sqlalchemy import paginate
+from sqlalchemy.orm import Session
+
+from lib.database import get_db
+from plataforma_web.v1.roles.models import Permiso
+from plataforma_web.v1.usuarios.authentications import get_current_active_user
+from plataforma_web.v1.usuarios.schemas import UsuarioInBD
+
+from .crud import get_materias, get_materia
+from .schemas import MateriaOut
+
+router = APIRouter()
+
+
+@router.get("", response_model=LimitOffsetPage[MateriaOut])
+async def list_paginate(
+    current_user: UsuarioInBD = Depends(get_current_active_user),
+    db: Session = Depends(get_db),
+):
+    """Listado paginado de materias"""
+    if not current_user.permissions & Permiso.VER_CATALOGOS == Permiso.VER_CATALOGOS:
+        raise HTTPException(status_code=403, detail="Forbidden (no tiene permiso).")
+    return paginate(get_materias(db))
+
+
+@router.get("/{materia_id}", response_model=MateriaOut)
+async def detail_from_id(
+    materia_id: int,
+    current_user: UsuarioInBD = Depends(get_current_active_user),
+    db: Session = Depends(get_db),
+):
+    """Detalle de una materia a partir de su id"""
+    if not current_user.permissions & Permiso.VER_CATALOGOS == Permiso.VER_CATALOGOS:
+        raise HTTPException(status_code=403, detail="Forbidden (no tiene permiso).")
+    materia = get_materia(db, materia_id)
+    return MateriaOut(
+        id=materia.id,
+        nombre=materia.nombre,
+    )
