+++ plataforma_web/v1/roles/schemas.py
@@ -1,0 +1,15 @@
+"""
+Roles v1.0, esquemas
+"""
+from pydantic import BaseModel
+
+
+class RolOut(BaseModel):
+    """Esquema para entregar rol"""
+
+    id: int
+    nombre: str
+
+    class Config:
+        """SQLAlchemy config"""
+        orm_mode = True
