+++ plataforma_web/v1/usuarios/paths.py
@@ -1,0 +1,48 @@
+"""
+Usuarios v1.0, rutas (paths)
+"""
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi_pagination import LimitOffsetPage
+from fastapi_pagination.ext.sqlalchemy import paginate
+from sqlalchemy.orm import Session
+
+from lib.database import get_db
+from plataforma_web.v1.roles.models import Permiso
+from .authentications import get_current_active_user
+from .crud import get_usuarios, get_usuario
+from .schemas import UsuarioOut, UsuarioInBD
+
+router = APIRouter()
+
+
+@router.get("", response_model=LimitOffsetPage[UsuarioOut])
+async def list_paginate(
+    autoridad_id: int = None,
+    current_user: UsuarioInBD = Depends(get_current_active_user),
+    db: Session = Depends(get_db),
+):
+    """Listado paginado de usuarios"""
+    if not current_user.permissions & Permiso.VER_CUENTAS == Permiso.VER_CUENTAS:
+        raise HTTPException(status_code=403, detail="Forbidden (no tiene permiso).")
+    return paginate(get_usuarios(db, autoridad_id=autoridad_id))
+
+
+@router.get("/{usuario_id}", response_model=UsuarioOut)
+async def detail_from_id(
+    usuario_id: int,
+    current_user: UsuarioInBD = Depends(get_current_active_user),
+    db: Session = Depends(get_db),
+):
+    """Detalle de un usuario a partir de su id"""
+    if not current_user.permissions & Permiso.VER_CUENTAS == Permiso.VER_CUENTAS:
+        raise HTTPException(status_code=403, detail="Forbidden (no tiene permiso).")
+    usuario = get_usuario(db, usuario_id=usuario_id)
+    return UsuarioOut(
+        id=usuario.id,
+        rol_id=usuario.rol_id,
+        autoridad_id=usuario.autoridad_id,
+        email=usuario.email,
+        nombres=usuario.nombres,
+        apellido_paterno=usuario.apellido_paterno,
+        apellido_materno=usuario.apellido_materno,
+    )
